+++ commandManager/app/Application/Port/In/Task/CreateTask/Service.go
@@ -1,10 +1,19 @@
 package CreateTask
 
-import "github.com/Enrikerf/pfm/commandManager/app/Domain/Model"
+import (
+	TaskOutPort "github.com/Enrikerf/pfm/commandManager/app/Application/Port/Out/Database/Task"
+	TaskDomain "github.com/Enrikerf/pfm/commandManager/app/Domain/Model/Task"
+)
 
-type Service struct{}
+type Service struct{
+	savePort TaskOutPort.SavePort
+}
 
-func (service Service) Create(command Command) (Model.Task, error) {
-	var task, _ = Model.NewTask(command.Host, command.Port, command.Command, command.Mode)
-	return task, nil
+func (service Service) Create(command Command) (TaskDomain.Task, error) {
+	var task, _ = TaskDomain.NewTask(command.Host, command.Port, command.Command, command.Mode)
+	newTask, err := service.savePort.Save(task)
+	if err != nil {
+		return TaskDomain.Task{}, err
+	}
+	return newTask, nil
 }
