+++ commandManager/go.mod
@@ -3,8 +3,10 @@ module github.com/Enrikerf/pfm/commandManager
 go 1.15
 
 require (
+	github.com/google/gops v0.3.22 // indirect
 	github.com/google/uuid v1.3.0
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.7.0
+	github.com/stretchr/testify v1.7.0
 	google.golang.org/grpc v1.42.0
 	google.golang.org/grpc/cmd/protoc-gen-go-grpc v1.1.0
 	google.golang.org/protobuf v1.27.1
